Rework entrypoints and subpackage __*__ files for new setup
