+++ setup.py
@@ -47,11 +47,17 @@ setup(
     packages=find_packages(),
 
     entry_points={'console_scripts': [
-            'hyp3_gamma = hyp3_gamma.__main__:entry',
-            'rtc =  hyp3_gamma.__main__:rtc',
-            'check_coreg.py = hyp3_gamma.rtc.check_coreg:main',
-            'rtc_sentinel.py = hyp3_gamma.rtc.rtc_sentinel:main',
-            'smooth_dem_tiles.py = hyp3_gamma.rtc.smoothem:main',
+        'hyp3_gamma = hyp3_gamma.__main__:entry',
+        'rtc = hyp3_gamma.__main__:rtc',
+        'check_coreg.py = hyp3_gamma.rtc.check_coreg:main',
+        'rtc_sentinel.py = hyp3_gamma.rtc.rtc_sentinel:main',
+        'smooth_dem_tiles.py = hyp3_gamma.rtc.smoothem:main',
+        'geocode = hyp3_gamma.__main__:geocode',
+        'geocode_sentinel.py = hyp3_gamma.geocode.sentinel:main',
+        'insar = hyp3_gamma.__main__:insar',
+        'ifm_sentinel.py = hyp3_gamma.insar.ifm_sentinel:main',
+        'interf_pwr_s1_lt_tops_proc.py = hyp3_gamma.insar.interf_pwr_s1_lt_tops_proc:main',
+        'unwrapping_geocoding.py = hyp3_gamma.insar.unwrapping_geocoding:main',
         ]
     },
